Refactoring the package name and method signatures
