+++ ekisandroidkeystoresignatureprovider/src/androidTest/java/ekis/PBTX/EkisAndroidKeyStoreSignatureProviderInstrumentedTest.kt
@@ -1,18 +1,11 @@
-package ekis.pbtxjavaandroidkeystoresignatureprovider
+package ekis.PBTX
 
-import android.util.Log
 import androidx.test.runner.AndroidJUnit4
-import com.google.crypto.tink.subtle.EllipticCurves
-import org.junit.Assert
+import ekis.PBTX.PbtxEkis.Companion.toHexString
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.ExpectedException
 import org.junit.runner.RunWith
-import java.lang.Exception
-import java.security.KeyFactory
-import java.security.KeyStore
-import java.security.interfaces.ECPublicKey
-import java.security.spec.X509EncodedKeySpec
 
 /**
  * Test class for [ekisAndroidKeyStoreSignatureProvider]
@@ -48,11 +41,40 @@ class EkisAndroidKeyStoreSignatureProviderInstrumentedTest {
 
         PbtxEkis.createKey(TEST_CONST_TEST_KEY_NAME)
 
+
+    }
+
+    @Test
+    fun deleteKeys() {
+
+        // Use the key that was just added to the keystore to sign a transaction.
+
+        PbtxEkis.deleteKey(TEST_CONST_TEST_KEY_NAME)
+
+
+    }
+
+    @Test
+    fun listKeys() {
+
+        // Use the key that was just added to the keystore to sign a transaction.
+        var KeyList = PbtxEkis.listKeys()
+
+        KeyList.forEach() {
+            System.out.print("Key " + it.key.toHexString());
+        }
+
+    }
+
+    @Test
+    fun signData() {
+
         var mPublicKeys = PbtxEkis.signData(
                 "0102030405060708090a0b0c0d0e0f".toByteArray(),
                 TEST_CONST_TEST_KEY_NAME
         )
 
     }
+
 
 }
