+++ ekisandroidkeystoresignatureprovider/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="ekis.pbtxjavaandroidkeystoresignatureprovider"/>
+<manifest package="ekis.PBTX"/>
