+++ ekisandroidkeystoresignatureprovider/src/main/java/ekis/PBTX/PbtxEkis.kt
@@ -1,19 +1,18 @@
-package ekis.pbtxjavaandroidkeystoresignatureprovider
+package ekis.PBTX
 
 import android.security.keystore.KeyGenParameterSpec
 import android.security.keystore.KeyProperties
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.AndroidKeyStoreDeleteError
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.AndroidKeyStoreSigningError
+import ekis.PBTX.errors.AndroidKeyStoreDeleteError
+import ekis.PBTX.errors.AndroidKeyStoreSigningError
 import com.google.crypto.tink.subtle.EllipticCurves
-import ekis.pbtxjavaandroidkeystoresignatureprovider.Model.KeyModel
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.*
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.ErrorString.Companion.DELETE_KEY_KEYSTORE_GENERIC_ERROR
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.ErrorString.Companion.GENERATE_KEY_ECGEN_MUST_USE_SECP256R1
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.ErrorString.Companion.GENERATE_KEY_KEYGENSPEC_MUST_USE_EC
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.ErrorString.Companion.GENERATE_KEY_MUST_HAS_PURPOSE_SIGN
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.ErrorString.Companion.QUERY_ANDROID_KEYSTORE_GENERIC_ERROR
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.InvalidKeyGenParameter
-import ekis.pbtxjavaandroidkeystoresignatureprovider.errors.QueryAndroidKeyStoreError
+import ekis.PBTX.Model.KeyModel
+import ekis.PBTX.errors.ErrorString.Companion.DELETE_KEY_KEYSTORE_GENERIC_ERROR
+import ekis.PBTX.errors.ErrorString.Companion.GENERATE_KEY_ECGEN_MUST_USE_SECP256R1
+import ekis.PBTX.errors.ErrorString.Companion.GENERATE_KEY_KEYGENSPEC_MUST_USE_EC
+import ekis.PBTX.errors.ErrorString.Companion.GENERATE_KEY_MUST_HAS_PURPOSE_SIGN
+import ekis.PBTX.errors.ErrorString.Companion.QUERY_ANDROID_KEYSTORE_GENERIC_ERROR
+import ekis.PBTX.errors.InvalidKeyGenParameter
+import ekis.PBTX.errors.QueryAndroidKeyStoreError
 import java.security.*
 import java.security.interfaces.ECPublicKey
 import java.security.spec.ECGenParameterSpec
@@ -41,8 +40,8 @@ class PbtxEkis {
          */
         @JvmStatic
         private fun generateAndroidKeyStoreKey(
-            keyGenParameterSpec: KeyGenParameterSpec,
-            alias: String
+                keyGenParameterSpec: KeyGenParameterSpec,
+                alias: String
         ): ByteArray {
             // Parameter Spec must include PURPOSE_SIGN
             if (KeyProperties.PURPOSE_SIGN and keyGenParameterSpec.purposes != KeyProperties.PURPOSE_SIGN) {
@@ -60,7 +59,7 @@ class PbtxEkis {
             }
 
             val kpg: KeyPairGenerator =
-                KeyPairGenerator.getInstance(KeyProperties.KEY_ALGORITHM_EC, ANDROID_KEYSTORE)
+                    KeyPairGenerator.getInstance(KeyProperties.KEY_ALGORITHM_EC, ANDROID_KEYSTORE)
             kpg.initialize(keyGenParameterSpec)
             kpg.generateKeyPair()
 
@@ -73,12 +72,14 @@ class PbtxEkis {
          * Generate a new key inside AndroidKeyStore by the given [alias] and return the new key in bye[] format
          *
          * The given [alias] is the identity of the key. The new key will be generated with the Default [KeyGenParameterSpec] from the [generateDefaultKeyGenParameterSpecBuilder]
+         *
+         * @param alias : Alias of the key store
          */
         @JvmStatic
         fun createKey(alias: String) {
             // Create a default KeyGenParameterSpec
             val keyGenParameterSpec: KeyGenParameterSpec =
-                generateDefaultKeyGenParameterSpecBuilder(alias).build()
+                    generateDefaultKeyGenParameterSpecBuilder(alias).build()
 
             generateAndroidKeyStoreKey(keyGenParameterSpec, alias)
         }
@@ -86,10 +87,7 @@ class PbtxEkis {
 
         /**
          * Get all (SECP256R1) keys in EOS format from Android KeyStore
-         * @param alias String - the key's identity
-         * @param password KeyStore.ProtectionParameter? - the password to load all the keys
-         * @param loadStoreParameter KeyStore.LoadStoreParameter? - the KeyStore Parameter to load the KeyStore instance
-         * @return Array of public Keys 
+         * @return Array of public Keys
          */
 
         @Throws(QueryAndroidKeyStoreError::class)
@@ -99,7 +97,7 @@ class PbtxEkis {
 
             try {
                 val keyStore =
-                    getKeystore(null)
+                        getKeystore(null)
                 var aliasList = keyStore.aliases().toList()
                 aliasList.forEach() {
 
@@ -114,8 +112,8 @@ class PbtxEkis {
 
             return mList;
         }
-        
-        
+
+
         /**
          * Delete key in the Android KeyStore with matching the alias name.
          *
@@ -135,7 +133,6 @@ class PbtxEkis {
         }
 
 
-
         private fun getKeystore(loadStoreParameter: KeyStore.LoadStoreParameter?): KeyStore {
             return KeyStore.getInstance(ANDROID_KEYSTORE).apply { load(loadStoreParameter) }
 
@@ -143,20 +140,20 @@ class PbtxEkis {
 
 
         private fun getProtobufKey(
-            keyStore: KeyStore,
-            alias: String,
-            password: KeyStore.ProtectionParameter?
+                keyStore: KeyStore,
+                alias: String,
+                password: KeyStore.ProtectionParameter?
         ): KeyModel {
             val keyEntry = keyStore.getEntry(alias, password) as KeyStore.PrivateKeyEntry
             val ecPublicKey =
-                KeyFactory.getInstance(keyEntry.certificate.publicKey.algorithm).generatePublic(
-                    X509EncodedKeySpec(keyEntry.certificate.publicKey.encoded)
-                ) as ECPublicKey
+                    KeyFactory.getInstance(keyEntry.certificate.publicKey.algorithm).generatePublic(
+                            X509EncodedKeySpec(keyEntry.certificate.publicKey.encoded)
+                    ) as ECPublicKey
 
             val bytes = EllipticCurves.pointEncode(
-                EllipticCurves.CurveType.NIST_P256,
-                EllipticCurves.PointFormatType.COMPRESSED,
-                ecPublicKey.w
+                    EllipticCurves.CurveType.NIST_P256,
+                    EllipticCurves.PointFormatType.COMPRESSED,
+                    ecPublicKey.w
             )
 
             val i = 1
@@ -190,8 +187,8 @@ class PbtxEkis {
         @Throws(AndroidKeyStoreSigningError::class)
         @JvmStatic
         fun signData(
-            data: ByteArray,
-            alias: String
+                data: ByteArray,
+                alias: String
         ): ByteArray? {
             try {
                 var keyStore = getKeystore(null)
@@ -233,11 +230,11 @@ class PbtxEkis {
         @JvmStatic
         private fun generateDefaultKeyGenParameterSpecBuilder(alias: String): KeyGenParameterSpec.Builder {
             return KeyGenParameterSpec.Builder(
-                alias,
-                KeyProperties.PURPOSE_SIGN
+                    alias,
+                    KeyProperties.PURPOSE_SIGN
             )
-                .setDigests(KeyProperties.DIGEST_SHA256)
-                .setAlgorithmParameterSpec(ECGenParameterSpec(SECP256R1_CURVE_NAME))
+                    .setDigests(KeyProperties.DIGEST_SHA256)
+                    .setAlgorithmParameterSpec(ECGenParameterSpec(SECP256R1_CURVE_NAME))
         }
     }
