+++ ekisandroidkeystoresignatureprovider/src/main/java/ekis/PBTX/errors/InvalidKeyGenParameter.kt
@@ -1,4 +1,4 @@
-package ekis.pbtxjavaandroidkeystoresignatureprovider.errors
+package ekis.PBTX.errors
 
 class InvalidKeyGenParameter : Exception {
     constructor(message: String?) : super(message)
